% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{capture_r_output}
\alias{capture_r_output}
\title{Capture nicely formatted R outputs to send in slack}
\usage{
capture_r_output(print_object)
}
\arguments{
\item{print_object}{An object that has a print method() associated with it}
}
\value{
A string that is compatible with sending R outputs in slack messages
}
\description{
Function to capture R outputs to send them as slack messages
}
\examples{
df <- data.frame(a = 1:5, d = 96:100)
print_df <- capture_r_output(df)
param_list <- list(username = 'Name of bot', channel = '@chris.kelly', icon_emoji = ':call_me_hand:')
param_list$text <- print_df
write_to_slack(param_list = param_list)

}
\keyword{slack}
