% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{link}
\alias{link}
\title{Encode text as link}
\usage{
link(text_to_display = "Click here", link = "https://google.co.uk")
}
\arguments{
\item{text_to_display}{The text to display in the message}

\item{link}{The link to direct the user to}
}
\value{
A string that is compatible with writing links in slack messages
}
\description{
Shortcut to write syntax for encoding text as link
}
\examples{
paste0('Hi there! '
, link('Check out our link', 'https://google.co.uk'))
# returns [1] "Hi there! <https://google.co.uk|Check out our link>"

}
\keyword{slack}
