% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{payload}
\alias{payload}
\title{Helper function to write payload to slack message}
\usage{
payload(param_list)
}
\arguments{
\item{param_list}{list of keys and values to create json payload}
}
\description{
Helper function to write payload to slack message
}
