% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{write_to_slack}
\alias{write_to_slack}
\title{Encode text as link}
\usage{
write_to_slack(url = Sys.getenv("SLACK_WEBHOOK_URL"),
  param_list = list(channel = "@chris.kelly", text = "Hello world"))
}
\arguments{
\item{url}{String: The url of the webhook. By default, takes the SLACK_WEBHOOK_URL environment variable}

\item{param_list}{List: List of keys and values to feed into the json payload to be}
}
\value{
A string that is compatible with writing links in slack messages
}
\description{
Shortcut to write syntax for encoding text as link
}
\examples{
attachments <- list(fallback = paste0("Example link: ", link(text_to_display = "Click here", link = 'https://google.co.uk'))
                    , pretext = paste0("Example link: ", link(text_to_display = "Click here", link = 'https://google.co.uk'))
                    , color = "#d12626"
                    , footer = "Attachment footer"
                    , fields = list(
                        list(
                          title = "Title of the attachment"
                          , value = "Text in the attachment"
                          , short = "false"
                        )
                      )
                    )
param_list <- list(text = 'The text to display in the body of the message'
                   , username = 'The name to give the "bot" that sends the message'
                   , channel = "@chris.kelly"
                   , icon_emoji = ':ghost:'
                   , attachments = list(attachments))
write_to_slack(param_list = param_list)

}
\seealso{
\href{https://api.slack.com/docs/messages/builder}{Sdlack message builder}
}
\keyword{slack}
